+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Under development
 
+- `Parser`
+    - `#document` -- Updated to lazy parse the document.
+
 ## 1.5 _(January 31, 2017)_
 
 - Executables